Backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.12.4
django-cors-headers>=3.7.0
mysqlclient>=2.1.1
python-dotenv>=1.0.0
django-extensions>=3.1.3
faker >= 9.0.2
qrcode>=7.3.1
cloudinary >=1.26.0
